requirements.txt
# Stage 1 PDF ingestion
pymupdf>=1.24.0

# Optional: full-page OCR for routed pages (requires Tesseract on PATH)
Pillow>=10.0.0
pytesseract>=0.3.10
# Optional: table extraction backend for --table-text-engine pdflumber
pdfplumber>=0.11.0

# Optional: PaddleOCR image/table OCR backend (2.x and 3.x supported; Stage1 picks init args at runtime)
# Install one Paddle runtime matching your environment:
# - CPU: pip install paddlepaddle
# - GPU: pip install paddlepaddle-gpu  (match CUDA version per Paddle install docs)
paddleocr>=2.7.0

# Stage 2: OpenAI extraction (set OPENAI_API_KEY) unless --ground-truth-as-catalog
openai>=1.40.0

# RAG (Stage1 --build-rag-index) + optional LLM narrowing (Stage2/4/4b/4c/5 --use-rag)
numpy>=1.24.0
faiss-cpu>=1.8.0

# Stage5: embedded open-source graph DB export (mission_graph_kuzu/)
kuzu>=0.6.0
# Stage 2: default page_coverage.png
matplotlib>=3.8.0

# Stage 1 optional: supplemental markdown corpus for Stage 2 recall boost
pymupdf4llm>=0.0.17

# Stage 1 optional: --export-reading-order pdf|docx (md needs no extra deps)
reportlab>=4.0.0
python-docx>=1.1.0
